requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
gitpython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
libclang==18.1.1
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.5.3
namex==0.1.0
narwhals==2.1.0
numpy==2.1.3
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==5.29.5
pyarrow==21.0.0
pydeck==0.9.1
pygments==2.19.2
pymongo==4.14.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
tornado==6.5.2
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
